Implement authorization checks in Order and User controllers; add createSecureOrder method in OrderService for enhanced order validation

+++ project-grupo-6/src/main/java/com/example/projectgrupo6/controllers/rest/OrderRestController.java
@@ -3,6 +3,7 @@ package com.example.projectgrupo6.controllers.rest;
 import com.example.projectgrupo6.domain.*;
 import com.example.projectgrupo6.dto.CartDTO;
 import com.example.projectgrupo6.dto.OrderDTO;
+import com.example.projectgrupo6.dto.UserDTO;
 import com.example.projectgrupo6.dto.basicDtos.CartItemBasicDTO;
 import com.example.projectgrupo6.dto.basicDtos.OrderBasicDTO;
 import com.example.projectgrupo6.dto.basicDtos.OrderItemBasicDTO;
@@ -103,39 +104,35 @@ public class OrderRestController {
     //POST
     //Order
     //Add validation (??)
+    //Validation is implemented in the orderService ;)
     @PostMapping("/user/{id}")
     public ResponseEntity<OrderBasicDTO> createOrder(@PathVariable long id, @RequestBody List<CartItemBasicDTO> itemBasicDTOS, HttpServletRequest request) {
+        
         if(!userService.isAuthorized(id, request)){
-            throw new IllegalArgumentException("Acceso denegado: No puedes acceder al item de la orden de otro usuario");
+            throw new IllegalArgumentException("Acceso denegado: No puedes crear una orden en nombre de otro usuario");
         }
-        
-        long currentUserId = userService.getCurrentUserId(request);
-        
-        if (userService.getById(currentUserId).isPresent()) {
-            List<CartItem> items = cartItemMapper.toDomainFromBasics(itemBasicDTOS);
-            //creates order
-            Order order = orderService.createOrderFromCart(userService.findById(currentUserId).get(), items, cartService.getCartTotal(currentUserId));
-            //deletes cart
-            for (CartItem item : items) {
-                cartService.removeItemFromCart(currentUserId, item.getId());
-            }
 
-            URI location = fromCurrentRequest().path("/{id}").buildAndExpand(order.getId()).toUri();;
-            return ResponseEntity.ok(orderMapper.toBasicDTO(order));
-        } else {
-            throw new NoSuchElementException();
+        long currentUserId = userService.getCurrentUserId(request);
+        User user = userService.getById(currentUserId)
+                .orElseThrow(() -> new NoSuchElementException("Usuario no encontrado"));
+
+        Order order = orderService.createSecureOrder(user, itemBasicDTOS);
+
+        for (CartItemBasicDTO dtoItem : itemBasicDTOS) {
+            cartService.removeItemFromCart(currentUserId, dtoItem.id());
         }
+
+        URI location = fromCurrentRequest().path("/{id}").buildAndExpand(order.getId()).toUri();
+        return ResponseEntity.created(location).body(orderMapper.toBasicDTO(order));
     }
 
     //PUT
     //Order
     //Add validation (???) -> so the user can't set the price to less (?)
+    //JUST ADMIN SHOULD BE ABLE TO UPDATE AN ORDER, SOOOO, NO VALIDATION NEEDED, THE ADMIN CAN DO WHATEVER HE WANTS WITH THE ORDER (???)
+    //Validation is implemented on WebSecurityConfig
     @PutMapping("/{ordId}/user/{id}")
     public OrderBasicDTO updateOrder(@PathVariable long ordId, @PathVariable long id, @RequestBody OrderBasicDTO orderBasicDTO, HttpServletRequest request) {
-        if(!userService.isAuthorized(id, request)){
-            throw new IllegalArgumentException("Acceso denegado: No puedes acceder a la orden de otro usuario");
-        }
-        
         if (!userService.getById(id).isPresent()) {
             throw new NoSuchElementException();
         }
@@ -148,7 +145,11 @@ public class OrderRestController {
     //DELETE
     //Order
     @DeleteMapping("/{ordId}/user/{id}")
-    public ResponseEntity<OrderDTO> deleteOrder(@PathVariable long ordId, @PathVariable long id) {
+    public ResponseEntity<OrderDTO> deleteOrder(@PathVariable long ordId, @PathVariable long id, HttpServletRequest request) {
+        if (!userService.isAuthorized(id, request)) {
+            throw new IllegalArgumentException("Acceso denegado: No puedes borrar el pedido de otro usuario");
+        }
+
         if (userService.getById(id).isPresent()) {
             OrderDTO order = orderMapper.toDTO(orderService.findById(ordId).orElseThrow());
             orderService.deleteById(ordId);
@@ -162,7 +163,15 @@ public class OrderRestController {
     //Create
     @PostMapping("/{ordId}/file")
     public ResponseEntity<Void> uploadFile(@PathVariable long ordId,
-                                           @RequestParam MultipartFile file) throws IOException {
+                                           @RequestParam MultipartFile file,
+                                           HttpServletRequest request) throws IOException {
+
+        
+                                            
+        long currentUserId = userService.getCurrentUserId(request);
+        if (!userService.isAuthorized(currentUserId, request)) {
+            throw new IllegalArgumentException("Acceso denegado: No puedes crear el pedido de otro usuario");
+        }
 
         orderService.addFileToOrder(ordId, file);
         return ResponseEntity.ok().build();
@@ -170,9 +179,14 @@ public class OrderRestController {
 
     //Get
     @GetMapping("/{ordId}/file")
-    public ResponseEntity<Resource> getFile(@PathVariable long ordId) throws IOException {
+    public ResponseEntity<Resource> getFile(@PathVariable long ordId, HttpServletRequest request) throws IOException {
         Order order = orderService.findById(ordId)
                 .orElseThrow();
+
+        long ownerId = order.getUser().getId();
+        if (!userService.isAuthorized(ownerId, request)) {
+            throw new IllegalArgumentException("Acceso denegado: No puedes descargar la factura de otro usuario");
+        }
 
         if (order.getFilePath() == null) {
             return ResponseEntity.notFound().build();
